error delet s3

+++ src/services/Profile/EditDesignProfileService.ts
@@ -51,8 +51,9 @@ class EditDesignProfileService {
 
         if (background_image) {
             const s3Storage = new S3Storage()
-
-            await s3Storage.deleteFile(getProfile["background_image"])
+            if (getProfile["background_image"]) {
+                await s3Storage.deleteFile(getProfile["background_image"])
+            }
             const upload = await s3Storage.saveFile(background_image)
 
             data["background_image"] = upload
